PhysState Detection and Anti Double Jump
By understanding if the character is in air or standing on the ground
we get to prevent double jump.
Also just theoretically when we have water as a trigger, the Physical
state will switch to water. What we do about it is a different story.
Also added Tags class. It’s conviniet to have them in a separate class
as variable so that auto-complitipn works and we have one place to
manage tags.